backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
python-multipart==0.0.9
google-generativeai==0.8.0
python-dotenv==1.0.1
pymongo==4.6.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
